split namespace InstanceSerialization into MsgPackSerialization and PhpDocParsing
