+++ src/MsgPackSerialization/FieldMetadata.php
@@ -1,9 +1,11 @@
 <?php
 // Compiler for PHP (aka KPHP) polyfills
-// Copyright (c) 2020 LLC «V Kontakte»
+// Copyright (c) 2020 LLC ï¿½V Kontakteï¿½
 // Distributed under the GPL v3 License, see LICENSE.notice.txt
 
-namespace KPHP\InstanceSerialization;
+namespace KPHP\MsgPackSerialization;
+
+use KPHP\PhpDocParsing\PhpDocType;
 
 class FieldMetadata {
   /** @var int */
